+++ src/Factory/LineItemFactoryInterface.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Sylius\InvoicingPlugin\Factory;
 
+use Sylius\Component\Core\Model\ProductVariantInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 use Sylius\InvoicingPlugin\Entity\LineItemInterface;
 use Sylius\Component\Core\Model\OrderItemInterface;
@@ -30,6 +31,7 @@ interface LineItemFactoryInterface extends FactoryInterface
         ?string $variantName = null,
         ?string $variantCode = null,
         ?string $taxRate = null,
-        ?string $taxRateCode = null
+        ?string $taxRateCode = null,
+        ?ProductVariantInterface $variant = null
     ): LineItemInterface;
 }