+++ src/Generator/InvoicePdfFileGenerator.php
@@ -31,6 +31,10 @@ final class InvoicePdfFileGenerator implements InvoicePdfFileGeneratorInterface
 
     private string $template;
 
+    private string $headerTemplate;
+
+    private string $footerTemplate;
+
     private string $invoiceLogoPath;
 
     public function __construct(
@@ -39,6 +43,8 @@ final class InvoicePdfFileGenerator implements InvoicePdfFileGeneratorInterface
         FileLocatorInterface $fileLocator,
         InvoiceFileNameGeneratorInterface $invoiceFileNameGenerator,
         string $template,
+        string $headerTemplate,
+        string $footerTemplate,
         string $invoiceLogoPath
     ) {
         $this->templatingEngine = $templatingEngine;
@@ -46,6 +52,8 @@ final class InvoicePdfFileGenerator implements InvoicePdfFileGeneratorInterface
         $this->fileLocator = $fileLocator;
         $this->invoiceFileNameGenerator = $invoiceFileNameGenerator;
         $this->template = $template;
+        $this->headerTemplate = $headerTemplate;
+        $this->footerTemplate = $footerTemplate;
         $this->invoiceLogoPath = $invoiceLogoPath;
     }
 
@@ -53,12 +61,33 @@ final class InvoicePdfFileGenerator implements InvoicePdfFileGeneratorInterface
     {
         $filename = $this->invoiceFileNameGenerator->generateForPdf($invoice);
 
+        $options = [
+            'margin-top' => 20,
+            'margin-right' => 20,
+            'margin-bottom' => 20,
+            'margin-left' => 20,
+            // Header html
+            'header-html' => $this->templatingEngine->render($this->headerTemplate, [
+                'invoice' => $invoice,
+                'channel' => $invoice->channel(),
+                'invoiceLogoPath' => $this->fileLocator->locate($this->invoiceLogoPath),
+            ]),
+            // Footer html
+            'footer-html' => $this->templatingEngine->render($this->footerTemplate, [
+                'invoice' => $invoice,
+                'channel' => $invoice->channel(),
+                'invoiceLogoPath' => $this->fileLocator->locate($this->invoiceLogoPath),
+            ]),
+        ];
+
         $pdf = $this->pdfGenerator->getOutputFromHtml(
             $this->templatingEngine->render($this->template, [
                 'invoice' => $invoice,
                 'channel' => $invoice->channel(),
                 'invoiceLogoPath' => $this->fileLocator->locate($this->invoiceLogoPath),
             ])
+            ,
+            $options
         );
 
         return new InvoicePdf($filename, $pdf);